+++ src/main/java/com/bank/api/dto/AccountResponse.java
@@ -2,19 +2,22 @@ package com.bank.api.dto;
 
 public class AccountResponse {
 
-    private int accountId;
+	private Long accountId;
     private double balance;
     private String currency;
     private String status;
 
-    public AccountResponse(int accountId, double balance, String currency, String status) {
+    public AccountResponse(Long accountId,
+            double balance,
+            String currency,
+            String status) {
         this.accountId = accountId;
         this.balance = balance;
         this.currency = currency;
         this.status = status;
     }
 
-    public int getAccountId() {
+    public Long getAccountId() {
         return accountId;
     }
